Started Files controller refactoring
