+++ src/components/Controllers/Files/Buttons/DeleteButton.tsx
@@ -1,0 +1,17 @@
+import * as React from 'react'
+import DefaultButton from '../../../Parts/Buttons/DefaultButton'
+import { useStore } from '../store'
+
+export default function DeleteButton() {
+  const path = useStore((state) => state.path)
+  const deleteFile = useStore((state) => state.deleteFile)
+  return (
+    <DefaultButton
+      disabled={!path}
+      variant="text"
+      label="Delete"
+      color="warning"
+      onClick={() => deleteFile()}
+    />
+  )
+}
