+++ src/components/Controllers/Files/Dialogs/CopyDialog.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react'
+
+export default function CopyDialog() {
+  return <div>copy</div>
+}
